backend/requirements.txt
fastapi
uvicorn
openai
git+https://github.com/openai/whisper.git
python-dotenv
ffmpeg-python
